+++ libvast/src/format/syslog.cpp
@@ -113,8 +113,9 @@ reader::read_impl(size_t max_events, size_t max_slice_size, consumer& f) {
         std::chrono::duration_cast<std::chrono::milliseconds>(read_timeout_));
     }
   };
+  auto start = std::chrono::steady_clock::now();
   for (size_t produced = 0; produced < max_events; timeout = next_line()) {
-    if (timeout) {
+    if (timeout || start + read_timeout_ < std::chrono::steady_clock::now()) {
       VAST_DEBUG(this, "reached input timeout at line", lines_->line_number());
       return finish(f, ec::timeout);
     }