+++ libvast/src/system/application.cpp
@@ -171,7 +171,8 @@ auto make_import_command() {
       .add<bool>("blocking,b", "block until the IMPORTER forwarded all data")
       .add<size_t>("max-events,n", "the maximum number of events to "
                                    "import")
-      .add<std::string>("read-timeout", "read timeout"));
+      .add<std::string>("read-timeout", "read timoeut after which data is "
+                                        "forwarded to the importer"));
   import_->add_subcommand("zeek", "imports Zeek logs from STDIN or file",
                           documentation::vast_import_zeek,
                           source_opts("?import.zeek"));